Move unmatched step marking into GherkinErrorMarker.
