+++ cucumber.eclipse.editor/src/main/java/cucumber/eclipse/editor/editors/Editor.java
@@ -1,23 +1,15 @@
 package cucumber.eclipse.editor.editors;
 
-import static cucumber.eclipse.editor.editors.FeatureFileUtil.getDocumentLanguage;
-import static cucumber.eclipse.editor.editors.FeatureFileUtil.getStepsInEncompassingProject;
 import gherkin.lexer.LexingError;
 import gherkin.parser.Parser;
 
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 import org.eclipse.core.resources.IFile;
-import org.eclipse.core.resources.IMarker;
-import org.eclipse.core.resources.IResource;
 import org.eclipse.core.runtime.CoreException;
-import org.eclipse.jface.text.BadLocationException;
-import org.eclipse.jface.text.FindReplaceDocumentAdapter;
 import org.eclipse.jface.text.IDocument;
-import org.eclipse.jface.text.IRegion;
 import org.eclipse.jface.text.Position;
 import org.eclipse.jface.text.source.Annotation;
 import org.eclipse.jface.text.source.ISourceViewer;
@@ -31,12 +23,8 @@ import org.eclipse.ui.IEditorInput;
 import org.eclipse.ui.IFileEditorInput;
 import org.eclipse.ui.contexts.IContextService;
 import org.eclipse.ui.editors.text.TextEditor;
-import org.eclipse.ui.texteditor.MarkerUtilities;
-
-import cucumber.eclipse.steps.integration.Step;
 
 public class Editor extends TextEditor {
-	private static final String UNMATCHED_STAP_ERROR_ID = "cucumber.eclipse.editor.editors.Editor.unmatchedsteperror";
 
 	private ColorManager colorManager;
 	private IEditorInput input;
@@ -149,7 +137,7 @@ public class Editor extends TextEditor {
 		IDocument doc = getDocumentProvider().getDocument(input);
 		IFileEditorInput fileEditorInput = (IFileEditorInput) input;
 		IFile featureFile = fileEditorInput.getFile();
-		GherkinErrorMarker marker = new GherkinErrorMarker(featureFile, doc);
+		GherkinErrorMarker marker = new GherkinErrorMarker(this, featureFile, doc);
 		marker.removeExistingMarkers();
 
 		Parser p = new Parser(marker, false);
@@ -157,60 +145,5 @@ public class Editor extends TextEditor {
 			p.parse(doc.get(), "", 0);
 		} catch (LexingError l) {
 		}
-
-		highlightUnmatchedSteps(featureFile, doc, marker.getSteps());
-	}
-
-	private void highlightUnmatchedSteps(IFile featureFile, IDocument doc,
-			List<gherkin.formatter.model.Step> list) {
-		deleteUnmatchedStepsMarkers(featureFile);
-		createMarkers(featureFile, doc, list);
-	}
-
-	private void createMarkers(IFile featureFile, IDocument doc,
-			List<gherkin.formatter.model.Step> list) {
-		for (gherkin.formatter.model.Step stepLine : list) {
-			String stepString = stepLine.getKeyword() + stepLine.getName();
-			Step step = new StepMatcher().matchSteps(getDocumentLanguage(this),
-					getStepsInEncompassingProject(featureFile), stepString);
-			if (step == null) {
-				try {
-					markUnmatchedStep(featureFile, doc, stepLine);
-				} catch (CoreException e) {
-					e.printStackTrace();
-				} catch (BadLocationException e) {
-					e.printStackTrace();
-				}
-			}
-		}
-	}
-
-	private void deleteUnmatchedStepsMarkers(IFile featureFile) {
-		try {
-			featureFile.deleteMarkers(UNMATCHED_STAP_ERROR_ID, true,
-					IResource.DEPTH_ZERO);
-		} catch (CoreException e1) {
-			e1.printStackTrace();
-		}
-	}
-
-	private void markUnmatchedStep(IFile featureFile, IDocument doc,
-			gherkin.formatter.model.Step stepLine) throws BadLocationException,
-			CoreException {
-		Map<String, Object> attributes = new HashMap<String, Object>();
-		attributes.put(IMarker.SEVERITY, IMarker.SEVERITY_WARNING);
-		attributes.put(IMarker.LINE_NUMBER, stepLine.getLine());
-		attributes.put(IMarker.MESSAGE,
-				"Step does not have a matching glue code.");
-
-		FindReplaceDocumentAdapter find = new FindReplaceDocumentAdapter(doc);
-		IRegion region = find.find(0, stepLine.getName(), true, true, false,
-				false);
-
-		attributes.put(IMarker.CHAR_START, region.getOffset());
-		attributes.put(IMarker.CHAR_END,
-				region.getOffset() + region.getLength());
-		MarkerUtilities.createMarker(featureFile, attributes,
-				UNMATCHED_STAP_ERROR_ID);
 	}
 }
